Fixing reward reinvestment spot price calculation and adding more tests

+++ contracts/global/Errors.sol
@@ -11,7 +11,7 @@ library Errors {
         int256 underlyingRedeemed,
         int256 underlyingCashRequiredToSettle
     );
-    error SlippageTooHigh(uint32 slippage, uint32 limit);
+    error SlippageTooHigh(uint256 slippage, uint32 limit);
     error InSettlementCoolDown(uint32 lastSettlementTimestamp, uint32 coolDownInMinutes);
     /// @notice settleVault called when there is no debt
     error SettlementNotRequired();
